Add files via upload
All of the files you need to make images with known storm dates and times and MASC snowflake data, for either a single 3-Panel (script_MRR_make_images_nomovie.m) or a sequence of images for movie-making.  These files can not be used on their own to update the webpage.